ms-clusterizacao/requirements.txt
# Databricks
databricks-sdk
databricks-sql-connector
databricks.sql

# Data Science
pandas
numpy

# Machine Learning
scikit-learn